refactor deriveKey into parent class. change key for cod #745
